admin_interneta_site/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
boto3==1.24.95
botocore==1.27.95
celery==5.2.7
celery-progress==0.1.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.1.2
gunicorn==20.1.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-s3direct==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jmespath==1.0.1
kombu==5.2.4
mysqlclient==2.1.1
opensearch-py==2.0.1
Pillow==9.2.0
prompt-toolkit==3.0.32
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
